+++ src/components/common/header/SearchButton.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import SearchSVG from '@/assets/images/search.svg';
+import SearchSVG from '@/assets/images/common/search.svg';
 import {scale} from 'react-native-size-matters';
 
 const SearchButton = (): React.JSX.Element => {
